Add files via upload
This function is a generalization of Ridge regularization method and the LASSO penalty. Performs SPARSE STATIS dual analysis, using a combination of LASSO and Ridge, on a set of data matrices. This function allows you to completely remove weak variables as with LASSO regularization or to contract them to zero as in Ridge